Added basic stanza parser.
Uses a push parser and generates a object heirarchy like Python's ElementTree.
Emits on receiving a stanza. Not meant for use with beginning a stream.

It also doesn't handle multiple children with the same node name as required
for <presence/> for example ( http://xmpp.org/rfcs/rfc3921.html#rfc.section.2.2.2.2 )