Implemented basic SASL authentication.
There are a few validation issues and error reporting is pathetic. There is a hard-coded
password right now. But this can be modified to use the account manager.
What it requires is for the session to restart the stream and have a boolean
deciding if it is authenticated or not