Added Stanza
Stanza is a simpler equivalent to creating a new xml Document with a quick way to build up a xml tree.
It isn't meant to be very flexible, but rather a light-weight way to quickly write XMPP messages and quickly
access the tree properties using the c(), a() functions. It will also provide the base for a simpler
form of declarative validation of incoming data rather than checking with conditions in each case