Modified stanza to hold child elements as a list.
Hashes won't work in our case since ordering of elements is also
important. The most obvious example is a formatted IM message like

<message ...>Hi there <strong>Nikhil</strong> one two</message>

Here we are concerned about the nested HTML and the order of text
fragments, so we cannot just take all text together. The modified
API also allows only the text or text + XML to be extracted.

Children access might be a bit slow, but in the majority of XMPP
messages, nested heirarchies are pretty rare and usually don't have
more than 2-3 children, so ideally the performance hit won't matter
much. It can always be rewritten later if required.